agcom2/requirements.txt
# Core dependencies
ulid-py==1.1.0

# REST API dependencies
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Monitoring (optional)
prometheus-fastapi-instrumentator==6.1.0

# Development
black==23.12.1
ruff==0.1.9
mypy==1.8.0
